builds basic table data structure from the file
At this point, we have a basic table that holds cells objects in a hash map.
No calculations of formulas yet, we just print the raw cell contents in stdout.

The idea is that the map keys will hold a String reference of the cell,
as "A1", "B2", etc. This will make it easy to search them and iterate,
so they can reference each other quite trivially.

Each cell will hold a lexer and a parser (to be implemented), so the
strings will be converted to an internal syntax tree and calculated.