completes the infrastructure for the cells to read references
Rust compiler makes it impossible to create circular references or data races, which introduces a big challenge on building data structures that are common in other languages. Because of the ownership model, referencing a shared pointer requires the knowledge of all types of Rust smart pointers, the subtle differences between them, and when each one is suitable. In the end, it was a RefCell wrapped in a Reference Counter, a case that should be avoided if possible, which I tried hard to achieve, but our program is one of such cases where a RefCell is required because we manage shared references to data that can be changed at runtime.